refactor(node): abstract child components
